feat: add logic pdf download
Took 2 hours 41 minutes

+++ packages/sanf/core/src/Modules/Financing/Dto/SimulationCalculationRequestDto.php
@@ -10,7 +10,7 @@ class SimulationCalculationRequestDto extends DataTransferObject
 {
     public ?int $financing_method_id;
     public ?int $financing_amount;
-    public ?float $down_payment_percentage;
+    public ?int $down_payment_percentage;
     public ?int $down_payment_amount;
     public ?int $tenor_in_month;
     public ?bool $is_send_email;
